Unique ids for service and requests
The app team is having difficulties identifying reviews, requests, and submissions without any kind of unique identifier. (e.g. there is a #1 request and #1 submission).

This commit changes the internal labor market counters to just one, to allow for unique ID's within a market. Not absolutely ideal for within the solidity, but having to use labor market address, requestId, submissionId, and user address to identify a review is weighing the db down.